pyproject.toml
[project]
name = "replan2eplus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Juliet Nwagwu Ume-Ezeoke", email = "jnwagwu@stanford.edu" }
]
requires-python = ">=3.13"
dependencies = [
    "eppy>=0.5.63",
    "geomeppy>=0.11.8",
    "ladybug-core>=0.44.19",
    "pyprojroot>=0.3.0",
    "rich>=14.1.0",
    "utils4plans>=0.1.6",
]

[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"
